+++ CHANGES.md
@@ -6,5 +6,6 @@
 + Added Flexible, Fixed modes.
 + Added identity label retraction/slide on url-bar focus.
 + Added preferences for modes, widths, identity label retraction.
++ Added ability in preferences to modify min/max-width of tabs.
 + Fixed graphical titlebar buttons overflow bugs.
 + Fixed inability to remove/move the added search button.